remove image
Implemented the function of removing associated images from the page. In addition, the image saving rules have been modified. Now only the resized image is saved, the original image will not be saved. Also, when there is information entered by the user on the page, if what is uploaded is not a picture, the entered information will not be cleared after the system reports an error.